Coroutine Guide Basic - update description (Basic1)
rename sample code
